Rename implementation and implement fallback
