+++ src/main/java/org/dita/dost/writer/AbstractXMLFilter.java
@@ -29,7 +29,7 @@ public abstract class AbstractXMLFilter extends XMLFilterImpl implements Abstrac
 
     protected DITAOTLogger logger;
     protected final XMLUtils xmlUtils = new XMLUtils();
-    Job job;
+    protected Job job;
     /** Absolute temporary directory URI to file being processed */
     protected URI currentFile;
     protected final Map<String, String> params = new HashMap<>();
