+++ src/test/resources/IncludeResolverTest/exp/include_xml.dita
