+++ src/test/resources/IncludeResolverTest/src/code.xml
