+++ src/test/resources/IncludeResolverTest/src/include_xml.dita
