+++ src/test/resources/IncludeResolverTest/src/include_xml_schema.dita
