+++ src/test/resources/IncludeResolverTest/src/schema.dtd
