+++ src/test/resources/IncludeResolverTest/src/schema.xml
