+++ src/test/resources/IncludeResolverTest/src/schema_invalid.xml
