+++ src/test/resources/IncludeResolverTest/src/utf-8.xml
